+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Next
+  - IndicatorBox can handle masks
   - added yml file to create test environment
   - LeastSquares fixed docstring and unified gradient code when out is passed or not.
   - Add compression to 8bit and 16bit to TIFFWriter